feat(docs): Enable MathJax physics package and clean up CSS
Enable the MathJax physics package to support advanced mathematical notation in documentation. This enhances the ability to render complex physics-related equations and symbols.

Additionally, this commit deprecates and removes the `api-autonav.css` file and its inclusion from `mkdocs.yaml.jinja`. This cleanup is performed because the `mkdocs-api-autonav` plugin is no longer actively used or supported in this template.

Minor stylistic improvements and browser compatibility adjustments are also included in `material.css`.